Move code to .cpp
